refactor: extract scroll-to-top context and hook to dedicated files
- Move ScrollToTopContext to src/contexts/scroll-to-top-context.tsx
- Move useScrollToTop hook to src/hooks/use-scroll-to-top.ts
- Update all imports across the codebase
- Improve code organization and maintainability
