+++ src/components/layout/index.ts
@@ -1,7 +1,7 @@
 // Wrappers
 export { ScreenWrapper } from './wrappers/screen-wrapper';
 export type { ScreenWrapperProps } from './wrappers/screen-wrapper';
-export { TabScreenWrapper, useScrollToTop } from './wrappers/tab-screen-wrapper';
+export { TabScreenWrapper } from './wrappers/tab-screen-wrapper';
 export type { TabScreenWrapperProps } from './wrappers/tab-screen-wrapper';
 export { ScrollableParallaxView } from './wrappers/scrollable-parallax-view';
